requirements.txt
Django==3.2.5
django-crispy-forms==1.12.0
django-phonenumber-field==5.2.0
phonenumbers==8.12.28
djangorestframework==3.12.4
django-filter